Use synthName-voiceID.dic for voice dict files
This converts dictionaries to use the ID rather than the name for the
file name. The downside of this is that for synths like SAPI5 or onecore
the filenames become quite long, and not really user facing. EG an ID
like the following SAPI5 one for English US David:
"HKEY_LOCAL_MACHINE\\SOFTWARE\\Microsoft\\Speech\\Voices\\Tokens\\TTS_MS_EN-US_DAVID_11.0"

We will hold off on this change until we have a more user friendly way of
managing the dictionaries.